AuthorizeAttribute and AuthorizeFilterAttribute merge on Asp.NET Core, supported on IDotvvmPresenter
