+++ src/DotVVM.Framework.Hosting.AspNetCore/Extensions/ApplicationBuilderExtensions.cs
@@ -37,7 +37,7 @@ namespace Microsoft.AspNetCore.Builder
 
             config.Debug = env.IsDevelopment();
             config.ApplicationPhysicalPath = applicationRootPath ?? env.ContentRootPath;
-            config.Runtime.GlobalFilters.Add(new AuthorizeFilterAttribute());
+            config.Runtime.GlobalFilters.Add(new AuthorizeAttribute());
 
             if (useErrorPages ?? env.IsDevelopment())
             {
