+++ src/DotVVM.Framework.Hosting.Owin/Dotvvm.Framework.Hosting.Owin.csproj
@@ -89,7 +89,6 @@
     <Compile Include="Extensions\DotvvmRequestContextExtensions.cs" />
     <Compile Include="Hosting\DotvvmAuthenticationHelper.cs" />
     <Compile Include="Runtime\Filters\AuthorizeAttribute.cs" />
-    <Compile Include="Runtime\Filters\NotAuthorizedAttribute.cs" />
     <Compile Include="Hosting\DotvvmCookieCollection.cs" />
     <Compile Include="Hosting\DotvvmHeaderCollection.cs" />
     <Compile Include="Hosting\DotvvmHttpContext.cs" />
